refactor : 베팅 패키지 추가
