+++ src/main/java/blackjack/domain/participants/betting/Betting.java
@@ -1,5 +1,7 @@
-package blackjack.domain.participants;
+package blackjack.domain.participants.betting;
 
+import blackjack.domain.participants.Player;
+import blackjack.domain.participants.Result;
 import java.util.Map;
 
 public class Betting {
