+++ src/main/java/blackjack/domain/participants/betting/BettingRepository.java
@@ -1,5 +1,6 @@
-package blackjack.domain.participants;
+package blackjack.domain.participants.betting;
 
+import blackjack.domain.participants.Player;
 import java.util.LinkedHashMap;
 import java.util.Map;
