+++ src/main/java/blackjack/view/OutputView.java
@@ -1,8 +1,8 @@
 package blackjack.view;
 
-import blackjack.domain.participants.Betting;
+import blackjack.domain.participants.betting.Betting;
 import blackjack.domain.participants.GameParticipant;
-import blackjack.domain.participants.Profit;
+import blackjack.domain.participants.betting.Profit;
 import blackjack.dto.ParticipantDto;
 import java.util.List;
 import java.util.stream.Collectors;
