+++ src/test/java/blackjack/domain/BettingTest.java
@@ -3,11 +3,11 @@ package blackjack.domain;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Rank;
 import blackjack.domain.card.Shape;
-import blackjack.domain.participants.Betting;
+import blackjack.domain.participants.betting.Betting;
 import blackjack.domain.participants.Hands;
 import blackjack.domain.participants.Name;
 import blackjack.domain.participants.Player;
-import blackjack.domain.participants.Profit;
+import blackjack.domain.participants.betting.Profit;
 import blackjack.domain.participants.Result;
 import java.util.List;
 import org.assertj.core.api.Assertions;
